feat: MYB-63 ip info logic changed
